Adding a very basic frame rate controller for the Test Runner project

+++ ModelViewer/Common/source/base/Renderer_Base.h
@@ -15,7 +15,7 @@ public:
 
     virtual Graphics::GraphicsError Initialize(API_Base *api) = 0;
     virtual Graphics::GraphicsError Finalize() = 0;
-    virtual Graphics::GraphicsError Update(f32 deltaTime) = 0;
+    virtual Graphics::GraphicsError Update(f64 deltaTime) = 0;
 
 };
